# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Car Service Station API",
        "description": "Description of API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api"
        }
    ],
    "paths": {
        "/schedules/import": {
            "post": {
                "tags": [
                    "schedules"
                ],
                "summary": "Импортирование расписания",
                "operationId": "import",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Import"
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
        "/stations": {
            "get": {
                "tags": [
                    "stations"
                ],
                "summary": "Список СТОА",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "stations"
                ],
                "summary": "Добавление новой СТОА",
                "operationId": "store",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Station"
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Station"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
        "/stations/{id}": {
            "get": {
                "tags": [
                    "stations"
                ],
                "summary": "Информация о СТОА",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "stations"
                ],
                "summary": "Редактирвоание информации о СТОА",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Station"
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Station"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Company not found"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "stations"
                ],
                "summary": "Удаление СТОА",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Company not found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Station": {
                "title": "СТОА",
                "description": "Станция Технического Обслуживания",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "name",
                        "type": "string",
                        "format": "string",
                        "example": "Название"
                    },
                    "address": {
                        "title": "address",
                        "description": "address",
                        "type": "string",
                        "format": "string",
                        "example": "Адресс"
                    },
                    "description": {
                        "title": "description",
                        "description": "description",
                        "type": "text",
                        "format": "text",
                        "example": "Описание"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "Station": {
                "description": "Station request",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/Station"
                        }
                    }
                }
            },
            "Import": {
                "description": "Schedule import",
                "required": true,
                "content": {
                    "multipart/form-data": {
                        "schema": {
                            "required": [
                                "file"
                            ],
                            "properties": {
                                "file": {
                                    "description": "file to upload",
                                    "type": "file",
                                    "format": "file"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "basic": {
                "type": "http",
                "description": "Use your email and api_key",
                "name": "Password Based",
                "in": "header",
                "scheme": "basic"
            }
        }
    }
}